Consolidate e2e wait/poll helpers onto testutil.Eventually
Move waitForUIContains/waitForAgentSessions/waitForNoAgentSessions/
waitForCond/assertAgentSessionsStayLive and friends from per-test files
into a shared non-_test internal/e2e/helpers.go, implemented on
testutil.Eventually plus a new testutil.Consistently for the
stays-true-for-duration assertions. Pure test refactor.
