+++ internal/testutil/wait.go
@@ -26,6 +26,20 @@ func Eventually(t *testing.T, timeout, interval time.Duration, cond func() bool,
 	}
 }
 
+// Consistently polls check for the full duration at the given interval,
+// failing the test the first time check returns a non-empty failure message.
+// It is the inverse of Eventually: the condition must hold the whole time.
+func Consistently(t *testing.T, duration, interval time.Duration, check func() string) {
+	t.Helper()
+	deadline := time.Now().Add(duration)
+	for time.Now().Before(deadline) {
+		if msg := check(); msg != "" {
+			t.Fatalf("%s", msg)
+		}
+		time.Sleep(interval)
+	}
+}
+
 // WaitForAtomic polls load until it reports a value >= want or timeout elapses,
 // failing the test on timeout. load typically reads an atomic counter.
 func WaitForAtomic[T cmp.Ordered](t *testing.T, load func() T, want T, timeout time.Duration) {
